+++ hooks/src/lib/index.ts
@@ -1,4 +1,4 @@
-export { default as useCardNumbers } from "./useCardNumbers";
-export { default as useExpiryDate } from "./useExpiryDate";
-export { default as useCvcNumber } from "./useCvcNumber";
-export { default as usePassword } from "./usePassword";
+export { default as useCardNumbers } from './useCardNumbers/useCardNumbers';
+export { default as useExpiryDate } from './useExpiryDate/useExpiryDate';
+export { default as useCvcNumber } from './useCvcNumber/useCvcNumber';
+export { default as usePassword } from './usePassword/usePassword';
