+++ hooks/src/lib/useCvcNumber/useCvcNumber.ts
@@ -1,10 +1,10 @@
 import { useState } from 'react';
-import { INPUT_RULE } from './constants/inputRule';
-import { ERROR_MESSAGE } from './constants/errorMessage';
-import { isEmpty, isValidLength, isNumber } from './utils/validate';
-import { isOverInputLength } from './utils/overInputLength';
-import { initialError } from './utils/initial';
-import { ErrorType } from './types/errorType';
+import { INPUT_RULE } from '../constants/inputRule';
+import { ERROR_MESSAGE } from '../constants/errorMessage';
+import { isEmpty, isValidLength, isNumber } from '../utils/validate';
+import { isOverInputLength } from '../utils/overInputLength';
+import { initialError } from '../utils/initial';
+import { ErrorType } from '../types/errorType';
 
 type ValitationResult = {
   cvc: string;
